+++ src/DrawableSuits/PluginInfo.cs
@@ -4,5 +4,5 @@ internal static class PluginInfo
 {
     public const string Guid = "com.toxcgang.drawablesuits";
     public const string Name = "DrawableSuits";
-    public const string Version = "0.5.17";
+    public const string Version = "0.5.18";
 }